refactor: move camera-ready artifact writers
